requirements.txt
pandas==2.1.4
matplotlib==3.8.2
scikit-learn==1.4.0
joblib==1.3.2
cloudpickle==3.1.0
xgboost==2.1.3
mlflow==2.18.0
apache-airflow==2.10.3
dvc==3.58.0
